+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Use string key in produced message
+- Add `MessageKey` type and module
+- [**BC**] Add `MessageToProduce` type and module and use it instead of a `string` message in all produce functions
+    - Remove functions
+        - `produceWithHeaders`
+        - `produceSingleWithHeaders`
+- [**BC**] Remove partition from a producer runtime information and use a partition from a delivery report
+- [**BC**] Remove function for consuming last message
+    - `consumeLast`
+    - `consumeLastMessage`
+- [**BC**] Remove partition from consumer runtime information and use a partition from consumed message
+- [**BC**] Add `Partition` field to consumed `Message`
+- Trace offset in both producing and consuming message
 
 ## 19.0.0 - 2022-01-06
 - [**BC**] Use net6.0